# src/main/java/com/attornatus/backenddevelopertest/entities/Endereco.java
package com.attornatus.backenddevelopertest.entities;

import jakarta.persistence.Entity;
import jakarta.persistence.Id;
import org.springframework.hateoas.RepresentationModel;

import java.util.Objects;

@Entity
public class Endereco extends RepresentationModel<Endereco> {

    @Id
    private String id;
    private String idPessoa;
    private String logradouro;
    private String cep;
    private int numero;
    private String cidade;
    private boolean principal;

    public Endereco() {

    }

    public Endereco(String idPessoa, String logradouro, String cep, int numero, String cidade, boolean principal) {
        this.idPessoa = idPessoa;
        this.logradouro = logradouro;
        this.cep = cep;
        this.numero = numero;
        this.cidade = cidade;
        this.principal = principal;
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getIdPessoa() {
        return idPessoa;
    }

    public void setIdPessoa(String idPessoa) {
        this.idPessoa = idPessoa;
    }

    public String getLogradouro() {
        return logradouro;
    }

    public void setLogradouro(String logradouro) {
        this.logradouro = logradouro;
    }

    public String getCep() {
        return cep;
    }

    public void setCep(String cep) {
        this.cep = cep;
    }

    public int getNumero() {
        return numero;
    }

    public void setNumero(int numero) {
        this.numero = numero;
    }

    public String getCidade() {
        return cidade;
    }

    public void setCidade(String cidade) {
        this.cidade = cidade;
    }

    public boolean isPrincipal() {
        return principal;
    }

    public void setPrincipal(boolean principal) {
        this.principal = principal;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Endereco endereco = (Endereco) o;
        return numero == endereco.numero && principal == endereco.principal && Objects.equals(logradouro, endereco.logradouro) && Objects.equals(cep, endereco.cep) && Objects.equals(cidade, endereco.cidade);
    }

    @Override
    public int hashCode() {
        return Objects.hash(logradouro, cep, numero, cidade, principal);
    }

    @Override
    public String toString() {
        return "Endereco{" +
                "logradouro='" + logradouro + '\'' +
                ", cep='" + cep + '\'' +
                ", numero=" + numero +
                ", cidade='" + cidade + '\'' +
                ", principal=" + principal +
                '}';
    }
}